services/researcher/requirements.txt
langchain
langchain-community
langchain-google-genai
tavily-python
kafka-python
pydantic
pydantic-settings
structlog
python-json-logger
openai
groq
google-generativeai
prometheus-client==0.19.0
psycopg2-binary
sqlalchemy
